[v5] [core] feat: restore Toaster.create() support
