+++ packages/core/src/components/index.ts
@@ -103,9 +103,10 @@ export { Tab, TabId, TabProps } from "./tabs/tab";
 export { Tabs, TabsProps } from "./tabs/tabs";
 export { Tag, TagProps } from "./tag/tag";
 export { TagInput, TagInputProps, TagInputAddMethod } from "./tag-input/tagInput";
-export { OverlayToaster, OverlayToasterProps } from "./toast/overlayToaster";
+export { OverlayToaster } from "./toast/overlayToaster";
+export type { OverlayToasterProps, ToasterPosition } from "./toast/overlayToasterProps";
 export { Toast, ToastProps } from "./toast/toast";
-export { Toaster, ToastOptions, ToasterPosition } from "./toast/toaster";
+export { Toaster, ToastOptions } from "./toast/toaster";
 export { TooltipProps, Tooltip } from "./tooltip/tooltip";
 export { Tree, TreeProps } from "./tree/tree";
 export { TreeNodeInfo, TreeEventHandler } from "./tree/treeTypes";
